feat: Add function to check if card is in frame
This commit adds a new function `check_if_card_in_frame` to the `helpers.py` file. The function takes a frame as input and checks if a card is present in the center of the frame. It does this by extracting a 50x50 pixel region from the center of the frame and calculating the mean value of the pixels. If the mean value is greater than 150, it indicates the presence of a card. The function returns `True` if a card is present and `False` otherwise.