refactor: Add package-mode flag to pyproject.toml
This commit adds the package-mode flag to the pyproject.toml file. By setting the package-mode flag to false, the build system will not create a distributable package during the build process. This is useful when the project is being used as a library and does not need to be distributed as a package.

Note: The recent user commits and recent repository commits did not provide any relevant information for the commit message.