refactor: Improve OCR performance by measuring execution time
This commit adds timing measurements to the OCR process in the `run_ocr` function. By calculating the time taken for OCR, we can evaluate the performance of the OCR algorithm and identify any potential bottlenecks. This information will be useful for optimizing the OCR process and improving overall system efficiency.